# api.yaml
openapi: 3.0.3
info:
  title: GoChat IM API
  description: |
    A comprehensive instant messaging API built with Go, providing real-time chat functionality,
    user management, friend relationships, group chats, and file uploads.

    ## Features
    - JWT-based authentication
    - Real-time messaging via WebSocket
    - User registration and profile management
    - Friend relationship management
    - Group chat functionality
    - File upload with deduplication
    - Online status tracking

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## WebSocket Connection
    Real-time messaging uses WebSocket at `/ws?token=<jwt-token>`

  version: "1.0.0"
  contact:
    name: GoChat API Support
    email: support@gochat.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.gochat.com/api/v1
    description: Production server

# Security schemes
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: "BAD_REQUEST"
        message:
          type: string
          description: Human readable error message
          example: "Invalid request data"
        details:
          type: string
          description: Additional error details
          example: "Field 'email' is required"
      required:
        - code
        - message

    # Success response schema
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          description: Success code (always 0)
          example: 0
        message:
          type: string
          description: Success message
          example: "success"
        data:
          type: object
          description: Response data

    # User model
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: User unique identifier
          example: 1
        phone:
          type: string
          description: User phone number (11 digits)
          pattern: '^\d{11}$'
          example: "13800138000"
        nickname:
          type: string
          description: User display name
          minLength: 2
          maxLength: 50
          example: "John Doe"
        avatar:
          type: string
          description: Avatar image URL
          example: "avatars/user123.png"
        gender:
          type: integer
          description: User gender (0=unset, 1=male, 2=female)
          enum: [0, 1, 2]
          example: 1
        signature:
          type: string
          description: Personal signature
          maxLength: 200
          example: "Hello, I'm using GoChat!"
        created_at:
          type: string
          format: date-time
          description: Account creation time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last profile update time
          example: "2023-06-01T12:30:00Z"
      required:
        - id
        - phone
        - nickname

    # Message model
...
paths:
...
  /user/upload-avatar:
    post:
      summary: Upload user avatar
      description: Upload a new avatar image for the current user
      operationId: uploadAvatar
      tags:
        - User Management
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
                  description: Avatar image file (JPG, PNG, GIF)
              required:
                - avatar
      responses:
        '200':
          description: Avatar uploaded successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          avatar_url:
                            type: string
                            description: URL of the uploaded avatar
                            example: "avatars/user123_avatar.png"
        '400':
          description: Invalid file or file size too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/search:
    get:
      summary: Search users
      description: Search for users by phone number or nickname
      operationId: searchUsers
      tags:
        - User Management
      security:
        - bearerAuth: []
      parameters:
        - name: keyword
          in: query
          required: true
          description: Search keyword (phone number or nickname)
          schema:
            type: string
            minLength: 1
          example: "john"
        - name: page
          in: query
          required: false
          description: Page number (default 1)
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: page_size
          in: query
          required: false
          description: Number of results per page (default 20, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 10
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          users:
                            type: array
                            items:
                              $ref: '#/components/schemas/User'
                          total:
                            type: integer
                            description: Total number of matching users
                            example: 5
                          page:
                            type: integer
                            description: Current page number
                            example: 1
                          page_size:
                            type: integer
                            description: Results per page
                            example: 10
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Friend management endpoints
...
  /friend/add:
    post:
      summary: Add friend
      description: Send a friend request or add a friend directly
      operationId: addFriend
      tags:
        - Friend Management
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                friend_id:
                  type: integer
                  format: int64
                  description: ID of the user to add as friend
                  example: 2
                message:
                  type: string
                  maxLength: 200
                  description: Optional friend request message
                  example: "Hi, let's be friends!"
              required:
                - friend_id
            examples:
              add_friend:
                summary: Add friend request
                value:
                  friend_id: 2
                  message: "Hi, I'd like to add you as a friend!"
      responses:
        '200':
          description: Friend added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid friend ID or already friends
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /friend/{id}:
    delete:
      summary: Remove friend
      description: Remove a user from friend list
      operationId: removeFriend
      tags:
        - Friend Management
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Friend user ID to remove
          schema:
            type: integer
            format: int64
          example: 2
      responses:
        '200':
          description: Friend removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid friend ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Friend relationship not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Group management endpoints
...
  /online/users:
    get:
      summary: Get online users
      description: Get list of currently online users (friends only)
      operationId: getOnlineUsers
      tags:
        - Online Status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Online users list retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          online_users:
                            type: array
                            items:
                              type: object
                              properties:
                                user_id:
                                  type: integer
                                  format: int64
                                  example: 2
                                nickname:
                                  type: string
                                  example: "John Doe"
                                avatar:
                                  type: string
                                  example: "avatars/user123.png"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /online/count:
    get:
      summary: Get online user count
      description: Get total number of currently online users
      operationId: getOnlineCount
      tags:
        - Online Status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Online count retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          count:
                            type: integer
                            description: Number of online users
                            example: 25
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # File upload endpoints
...
security:
  - bearerAuth: []